requirements/base.txt
Django>=5.0,<6.0
djangorestframework>=3.15
django-filter>=24.0
drf-spectacular>=0.28
django-allauth>=0.61
django-guardian>=2.4
rules>=3.3
django-waffle>=4.0
psycopg[binary]>=3.1
redis>=5.0
celery[redis]>=5.4
django-storages[boto3]>=1.14
boto3>=1.34
whitenoise>=6.6
sentry-sdk>=2.0
django-csp>=3.8
django-ratelimit>=4.1
django-axes>=6.5
dj-database-url>=2.1
environs>=11.0
orjson>=3.10
prometheus-client>=0.20
python-json-logger>=2.0
stripe>=11.0
gunicorn>=21.2
uvicorn>=0.30
